+++ lib/podcast.en.rss.ts
@@ -1,5 +1,5 @@
-import { getAllPosts } from '../lib/getAllPosts';
-import { podcastXml } from '../lib/podcast';
+import { getAllPosts } from './getAllPosts';
+import { podcastXml } from './podcast';
 
 export const get = async () => {
   const englishPosts = (await getAllPosts()).map((dual) => dual.en);
